Rename `render:collection` => `render:children
Resolves #1309

I think this is the correct naming strategy.  It is what the function is called, which will likely be made [public](https://github.com/marionettejs/backbone.marionette/issues/2695)

Addressing previous concerns..  the difference with the collectionview's children vs children views of a view, would be that views are not responsible for rendering their children in the same way that collectionview's are, so I don't think there will be confusion anymore than our current situation, being that a view doesn't currently throw `render:collection` even though it can certainly render a collection.