Refactor: Add default ES6 options
This is mostly, just another syntatic cleanup. I added ES6 default
options and _.result default options.

It's worth pointing out that in the process I found a couple places
where we were not cloning the prototype property, which left the
opportunity open to mutate it. This is *not* a desirable behavior.

Also, I believe `|| {}` is mostly removed from the codebase, except
places where we need to do some local property accessing for passed in
options.