Check options for replaceElement / allowMissingEl
Previously, the Region object was only able to find `replaceElement` and `allowMissingEl` options if they were passed as constructor params. Now, the options object passed to `show` is also valid.

If both are provided, `show` takes precedence as it seems to make sense to override default values with method options.