Backbone radio docs (#2949)
* Started writing some Radio Docs #2647

I’ve put an introduction to the radio documentation in place.

The Radio is a major area where I don’t have a lot of personal
experience, so please go through this stuff with a fine-tooth comb
before merging into `next`

* Added some simple channel examples #2647

Added a couple of examples for channels. I’ve moved the documentation
above the other sections because the channels are really needed to
understand the rest of the documentation.

* Changed the initial example to match @paulfalgout's suggestion #2647

* Added a little more detail and extra outlines for binding events/requests to Marionette.Object #2647

* More Radio docs #2647

Written up outlines for all the main Radio concepts, linked to
Backbone, Radio, etc. where necessary.

* Minor readability improvement #2647

* Added some use-cases for the Radio #2647

I’ve written a couple of examples of when you might like to use Request
vs Event to help guide readers in one direction or another.

* Radio isn't used in View #2647

Changed references of View to Object - View does not extend Object so
it isn’t integrated in.