Went through the region docs #2915 (#2953)
* Went through the region docs #2915

I’ve replaced all references to an Application with View where it makes
more sense now.

* Change references of selector to el #2915 #2490

Updated references to selector after discussion with @paulfalgout in
#2955

* Made some improvements #2490 #2915

I’ve made a few improvements to the documentation for regions. I think
this will need a bigger ticket to totally re-factor how this works.