buildChildView, removeChildView and introduce addChildView (#2981)
`_addChild` `_buildChildView` and `_removeChildView` were made private for v3.

I've personally run into issues with needing `buildChildView` to show different views with different models than in the collection itself.  (essentially it uses the model in the collection to know which model type to get and show with a view)

I've also heard feedback that `addChild` would be missed.  However `_addChild` is doing a lot, and using it seems overly complex when you simply need to add a view as a child to the collectionView.