Region needs to detach previous html if there is no view
If a region had html in it on it's first show, the new view would get appended to the preexisting html.

This fix also makes a convienent place to override the jQuery detach

Unfortunately throwing the missing el error breaks soo many edge-casey things such as removing a region from an unrendered view.  This is already fixing an edge case, so it is probably best for the moment not to introduce a error.