Style Fixes (#3075)
* Cleaning up Template docs #3066

Did a little bike shedding and cleaned up the TemplateCache docs to fit
in the style of the rest. I’ve not changed the text at all here.

* More tweaks #3066

Made a few little style fixes to the Marionette.Object documentation.
I’ve not changed the text at all here.

* Whizzed through the rest of the docs #3066

Went through the rest of the docs to make them consistent

* Made some of the changes and import adds #3066