Advanced event doc (#3061)
* Made a start on the event doc #2920

I’ve started splitting the event docs into their own section. It feels
like this topic could be involved enough to warrant it.

* Whacked together some event docs #2920

I’ve put in some basic information about events.

I’ll take the text out of the marionette.collectionview.md file related
to childEvents and put it here too - with a link to it in that section.

* Added missing events attribute #2920

I’ve added the events attribute documentation which must have been
removed - it has some more detail though I’ll have to run through the
docs to make sure I’m not duplicating it.

Also made a start on the event bubbling docs.

* Added the triggers docs #2920

Added the triggers docs and referenced it into the events documentation
which we’ll need to flesh out more.

* Wrote a simple section on event bubbling #2920

Threw in some details about event bubbling and the childview prefix.

I might use this as an opportunity to move some of @paulfalgout’s work
on `CollectionView` into the `marionette.events.md` file later.

* Added the childViewEvents code #2920

Took the code from `marionette.collectionview.md` for the
childViewEvents and extended it a little bit.

* Link to view/collection view lifecycle events #2920