Separate template docs (#3108)
* Started separating the template docs out #3103

Separated the basic template logic out from the View to shorten the
view docs and give everything a clear place to point to.

This will also give us more room to expand on topic of templates.

* Removed duplicate template code #3103

Removed the template docs from view to get it below 1000 lines and
tweaked the template docs.