Collection view Events (#3120)
* Removed the newlines

* Renamed to events.md #3052

* Started moving the events around a bit #3052

Playing with the structure of events so it reflects marionette.view.md
a little more.

* Started re-ordering events to match View #3052

The lifecycle events should mirror that of View. We could potentially
extract this documentation into their own section but we’ll take a view
after this is completed.

* Start destruction lifecycle #3052

* Updated the destruction lifecycle #3052

I’ve updated the destruction lifecycle to follow the order laid out in
the table.

The only major change that needs to happen from here is to copy the
View’s lifecycle layout and probably move these events into their own
collective document.

* Reordered the creation/destruction events #3052